+++ packages/abstract-scanner/lib/constants.ts
@@ -1,2 +1,2 @@
-export const Seven_Days_InSeconds = 7 * 24 * 60 * 60;
-export const One_Hour_InSeconds = 60 * 60;
+export const BLOCK_CLEANUP_THRESHOLD_DURATION = 24 * 60 * 60;
+export const BLOCK_TRIM = 10;