+++ packages/scanners/cardano-scanner/lib/interfaces/index.ts
@@ -1,15 +1,10 @@
+import { BlockTimeConfig } from '@rosen-bridge/abstract-scanner/dist/scanner/interfaces';
 import { DataSource } from '@rosen-bridge/extended-typeorm';
 
 interface OgmiosReconnectionConfig {
   initialDelay?: number;
   maxDelay?: number;
   maxAttempts?: number;
-}
-
-interface BlockTimeConfig {
-  blockTime: number;
-  blockAgeThreshold?: number;
-  blockTrimCountInRound?: number;
 }
 
 interface CardanoOgmiosConfig {