+++ packages/scanners/firo-scanner/lib/scanner/firoRpcScanner.ts
@@ -13,6 +13,7 @@ export class FiroRpcScanner extends GeneralScanner<FiroRpcTransaction> {
       config.blockTimeConfig,
       config.logger,
       config.suffix,
+      config.heightGap,
     );
   }
 }