+++ packages/abstract-observation-extractor/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rosen-bridge/abstract-observation-extractor
 
+## 1.0.7
+
+### Patch Changes
+
+- Update dependencies
+  - @rosen-bridge/abstract-extractor@3.2.0
+  - @rosen-bridge/scanner-interfaces@1.0.0
+
 ## 1.0.6
 
 ### Patch Changes