+++ packages/box-tracker/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rosen-bridge/box-tracker
 
+## 0.1.5
+
+### Patch Changes
+
+- Update dependencies
+  - @rosen-bridge/abstract-extractor@3.2.0
+  - @rosen-bridge/scanner-interfaces@1.0.0
+
 ## 0.1.4
 
 ### Patch Changes